requirements.txt
beautifulsoup4
openpyxl
pytest
python-dotenv
requests
